stop using slick, use pandoc directly
