+++ src/Blog/Content.hs
@@ -6,8 +6,9 @@ module Blog.Content
   , substitute
   ) where
 
+import Blog.Core
 import Blog.Item
-import qualified Blog.Path.Rel as Rel
+import qualified Blog.Path.Rel as RelPath
 import Blog.Prelude
 import Blog.Types
 import qualified Blog.Wikilinks as CL
@@ -19,40 +20,54 @@ import qualified Data.Aeson as Aeson
 import Data.Functor.Identity (runIdentity)
 import qualified Data.Text as T
 import qualified Development.Shake.Plus as SP
-import qualified Slick
-import qualified Slick.Pandoc as Slick
-import qualified Text.Mustache.Types as M
+import qualified Text.Mustache as M
+import qualified Text.Mustache.Compile as M
+import Text.Pandoc (def)
 import qualified Text.Pandoc as Pandoc
+import qualified Text.Pandoc.Highlighting as Pandoc
 import qualified Text.Pandoc.Walk as PW
 
-compileTemplate :: Path Rel.Source File -> Action M.Template
-compileTemplate = SP.liftAction . Slick.compileTemplate' . toFilePath <=< Rel.sourceToRel
+compileTemplate :: Path RelPath.Source File -> Action M.Template
+compileTemplate templatePath = do
+  relPath <- RelPath.sourceToRel templatePath
+  SP.needP [relPath]
+  liftIO (M.localAutomaticCompile $ toFilePath relPath) >>= \case
+    Right template -> do
+      SP.need (M.getPartials . M.ast $ template)
+      pure template
+    Left err -> putError $ "[compileTemplate] error" <> show err
 
 substitute :: (Aeson.ToJSON a) => M.Template -> a -> Text
-substitute template = Slick.substitute template . Aeson.toJSON
+substitute template = M.substitute template . Aeson.toJSON
 
 -- | Default Pandoc markdown reader options. Using other options may break some of the helpers.
 options :: Pandoc.ReaderOptions
 options =
-  Slick.defaultMarkdownOptions
+  def
     { Pandoc.readerExtensions =
         mconcat
           [ Pandoc.extensionsFromList
-              [ Pandoc.Ext_auto_identifiers -- todo: test
-              , Pandoc.Ext_fenced_code_attributes -- todo: test
-              , Pandoc.Ext_footnotes -- todo: test
-              , Pandoc.Ext_wikilinks_title_after_pipe -- needed for obsidian links
-              , Pandoc.Ext_yaml_metadata_block -- needed for metadata
+              [ Pandoc.Ext_auto_identifiers
+              , Pandoc.Ext_fenced_code_attributes
+              , Pandoc.Ext_footnotes
+              , Pandoc.Ext_wikilinks_title_after_pipe
+              , Pandoc.Ext_yaml_metadata_block
               ]
           , Pandoc.githubMarkdownExtensions
           ]
+    }
+
+htmlOptions :: Pandoc.WriterOptions
+htmlOptions =
+  def
+    { Pandoc.writerHighlightStyle = Just Pandoc.tango
     }
 
 -- | Exported for testing. Should not be used.
 markdownToMetaAndContent :: (MonadIO m) => Text -> m (Aeson.Value, [Pandoc.Block])
 markdownToMetaAndContent content = do
   let
-    writer = Pandoc.writeHtml5String Slick.defaultHtml5Options
+    writer = Pandoc.writeHtml5String htmlOptions
   (Pandoc.Pandoc meta' blocks) <- unPandocM $ Pandoc.readMarkdown options content
   meta <- flattenMeta writer meta'
   pure (meta, blocks)
@@ -78,19 +93,19 @@ genenerateHtmlWithFixedWikiLinks
   -> ItemKind
   -> t Pandoc.Block
   -> Action Aeson.Value
-genenerateHtmlWithFixedWikiLinks cache def blocks = do
+genenerateHtmlWithFixedWikiLinks cache kind blocks = do
   let
     (result, logs) =
       runIdentity
         . flip State.runStateT []
         . Except.runExceptT
-        $ PW.walkM (CL.transform cache def) (toList blocks)
+        $ PW.walkM (CL.transform cache kind) (toList blocks)
   traverse_ printLog logs
   case result of
     Left _ -> crashWith "[Wikilinks] unrecoverable error. Stopping."
     Right fixedBlocks -> do
       let
-        writer = Pandoc.writeHtml5String Slick.defaultHtml5Options
+        writer = Pandoc.writeHtml5String htmlOptions
         doc = Pandoc.Pandoc mempty $ toList fixedBlocks
       outText <- unPandocM $ writer doc
       pure $ Aeson.String outText
