+++ src/Blog/Engine.hs
@@ -1,20 +1,10 @@
 module Blog.Engine
   ( runEngine
-  , want
   , copyFile
   , generatePage
   , initItemsCache
   , writeFile
-  , putVerbose
-  , putInfo
-  , putWarn
-  , putError
   , removeOutput
-  , (%>)
-  , need
-  , need'
-  , needItems
-  , (~>)
   , withMetadataObject
   , addKey
   , takeBaseName
@@ -24,6 +14,7 @@ module Blog.Engine
 where
 
 import qualified Blog.Content as Content
+import Blog.Core
 import Blog.Item
 import Blog.Path.Rel as Rel
 import Blog.Prelude
@@ -38,20 +29,11 @@ import qualified Data.Aeson.KeyMap as KeyMap
 import qualified Data.Text as T
 import qualified Development.Shake as Shake
 import qualified Development.Shake.Plus as SP
-import qualified System.FilePath as U
 
 runEngine :: Shake.ShakeOptions -> Settings.Settings -> Rules () -> IO ()
 runEngine args settings rules =
   Shake.shakeArgs args
     $ SP.runShakePlus settings rules
-
--- | Will attempt to find the appropriate Rules that match the requested paths.
---
--- You can define rules using @'(%>)'@.
---
--- 't' can be list, Maybe, etc.
-want :: (Partial, Traversable t) => t (Path Rel.Output File) -> Rules ()
-want files = traverse outputToRel files >>= SP.wantP
 
 -- | Copy a file from a source-relative path to a destination-relative path.
 copyFile :: (Partial) => Path Rel.Source File -> Path Rel.Output File -> Action ()
@@ -178,74 +160,12 @@ writeFile templatePath path' content = do
     . Content.substitute template
     $ content
 
-putVerbose :: String -> Action ()
-putVerbose = SP.liftAction . Shake.putVerbose
-
-putInfo :: String -> Action ()
-putInfo = SP.liftAction . Shake.putInfo
-
-putWarn :: String -> Action ()
-putWarn = SP.liftAction . Shake.putWarn
-
--- | Logs an error message and also exits the program.
-putError :: (Partial) => String -> Action a
-putError err = do
-  SP.liftAction $ Shake.putError err
-  crashWith err
-
-quietly :: Action a -> Action a
-quietly act = SP.withRunInAction (\run -> run act)
-
 -- | Nukes the output directory.
 removeOutput :: Action ()
 removeOutput = do
   output <- asks Settings.output
   putInfo $ "[removeOutput] Nuking output " <> show output
   SP.removeFiles output ["//"]
-
--- | Defines the build rules for an output path.
--- Example:
---
--- @
---     "post/*.html" %> \path -> do
---       ...
---       -- must generate path by the end, otherwise an error will be raised
--- @
-(%>) :: Shake.FilePattern -> (Path Rel.Output File -> Action ()) -> Rules ()
-(%>) pat action = do
-  output <- asks Settings.output
-  (toFilePath output U.</> pat) SP.%> \path -> do
-    putInfo $ "[%>] Generating " <> show path
-    p <- stripProperPrefix output path
-    action (asOutput p)
-
--- | Declare dependencies between rules through paths.
-need :: (Traversable t) => t (Path Rel.Output File) -> Action ()
-need path = traverse outputToRel path >>= SP.needP
-
--- | Declare dependencies between rules through paths.
-need' :: [Shake.FilePattern] -> Action ()
-need' pat = do
-  source <- asks Settings.source
-  SP.getDirectoryFiles source pat >>= need . fmap asOutput
-
--- | Needs all items in a cache.
--- Assumes that the output path is @<ItemKind>/<itemId.html>@
-needItems :: (ItemKind, [Item]) -> Action ()
-needItems (itemKind, items) = do
-  itemDir <- fmap asOutput . parseRelDir . T.unpack . getItemKind $ itemKind
-  need
-    . fmap (itemDir </>)
-    <=< traverse (addExtension ".html")
-    <=< traverse parseRelFile
-    . fmap (T.unpack . getItemId . id)
-    $ items
-
--- | Declare a pseudo-rule using a name rather than a path.
--- This is required because @'(%>)'@ would crash the program if used with a fake name.
--- Can be used to create rules for cleaning or other auxiliary tasks.
-(~>) :: String -> Action () -> Rules ()
-(~>) = SP.phony
 
 -- | Given some key @"k"@ and JSON value @v@, creates a JSON object @{ "k": v }@.
 withMetadataObject :: (Aeson.ToJSON a) => String -> a -> Aeson.Value
